+++ services/scheduler/src/main/java/org/mule/service/scheduler/internal/NullScheduledFuture.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
+ * The software in this package is published under the terms of the CPAL v1.0
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+package org.mule.service.scheduler.internal;
+
+import java.util.concurrent.Delayed;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+
+/**
+ * No-op implementation of a {@link ScheduledFuture}.
+ *
+ * @since 4.0
+ */
+final class NullScheduledFuture<V> implements ScheduledFuture<V> {
+
+  public static final NullScheduledFuture<?> INSTANCE = new NullScheduledFuture<>();
+
+  private NullScheduledFuture() {
+    // Nothing to do
+  }
+
+  @Override
+  public long getDelay(TimeUnit unit) {
+    return 0;
+  }
+
+  @Override
+  public int compareTo(Delayed o) {
+    return 0;
+  }
+
+  @Override
+  public boolean cancel(boolean mayInterruptIfRunning) {
+    return false;
+  }
+
+  @Override
+  public boolean isCancelled() {
+    return false;
+  }
+
+  @Override
+  public boolean isDone() {
+    return false;
+  }
+
+  @Override
+  public V get() throws InterruptedException, ExecutionException {
+    return null;
+  }
+
+  @Override
+  public V get(long timeout, TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
+    return null;
+  }
+}
